refactor(web): keep day calendar focus helpers in day
