+++ packages/web/src/common/utils/event/event.util.ts
@@ -11,10 +11,8 @@ import dayjs, { type Dayjs } from "@core/util/date/dayjs";
 import { getUserId } from "@web/auth/compass/session/session.util";
 import { isBackendUnavailableError } from "@web/common/apis/util/backend-unavailable-error.util";
 import {
-  DATA_CALENDAR_TIMED_GRID_ROW,
   DATA_EVENT_ELEMENT_ID,
   ID_GRID_ALLDAY_ROW,
-  ID_GRID_COLUMNS_TIMED,
   ID_GRID_MAIN,
 } from "@web/common/constants/web.constants";
 import { type PartialMouseEvent } from "@web/common/types/util.types";
@@ -25,12 +23,6 @@ import {
 } from "@web/common/types/web.event.types";
 import { reloadLocation } from "@web/common/utils/browser/browser-navigation.util";
 import { createObjectIdString } from "@web/common/utils/id/object-id.util";
-import {
-  focusDayCalendarEventTarget,
-  getFirstVisibleDayCalendarEventTarget,
-  getFocusedDayCalendarEventTarget,
-  getHoveredDayCalendarEventTarget,
-} from "@web/views/Day/interaction/targeting/dayCalendarEventTargeting";
 
 export const gridEventDefaultPosition: Schema_GridEvent["position"] = {
   isOverlapping: false,
@@ -172,73 +164,6 @@ export const getCalendarEventElementFromGrid = (
 
   return timedEvent ?? allDaySection?.querySelector(selector) ?? null;
 };
-
-export function openEventFormCreateEvent() {
-  const domEvent = new MouseEvent("mousedown", {
-    bubbles: true,
-    button: 0,
-    buttons: 1,
-  });
-
-  const calendarSurface = getTimedGridCreateTarget();
-
-  if (!calendarSurface) return;
-
-  calendarSurface.dispatchEvent(domEvent);
-}
-
-const getTimedGridCreateTarget = () => {
-  const timedGrid = document.getElementById(ID_GRID_MAIN);
-  const timedRow = timedGrid?.querySelector<HTMLElement>(
-    `[${DATA_CALENDAR_TIMED_GRID_ROW}="true"]`,
-  );
-
-  if (timedRow) {
-    return timedRow;
-  }
-
-  const timedColumns = document.getElementById(ID_GRID_COLUMNS_TIMED);
-  const timedRows = timedColumns?.nextElementSibling;
-  const firstTimedRow = timedRows?.firstElementChild;
-
-  if (firstTimedRow instanceof HTMLElement) {
-    return firstTimedRow;
-  }
-
-  return timedGrid;
-};
-
-export function openEventFormEditEvent() {
-  const target =
-    getFocusedDayCalendarEventTarget() ??
-    getHoveredDayCalendarEventTarget() ??
-    getFirstVisibleDayCalendarEventTarget();
-
-  if (!target) {
-    return;
-  }
-
-  target.element.scrollIntoView({ block: "nearest" });
-  focusDayCalendarEventTarget(target);
-  target.element.dispatchEvent(
-    new KeyboardEvent("keydown", {
-      bubbles: true,
-      cancelable: true,
-      key: "Enter",
-    }),
-  );
-}
-
-export function focusFirstDayCalendarEvent() {
-  const target = getFirstVisibleDayCalendarEventTarget();
-
-  if (!target) {
-    return;
-  }
-
-  target.element.scrollIntoView({ block: "nearest" });
-  focusDayCalendarEventTarget(target);
-}
 
 export const getMonthListLabel = (start: Dayjs) => {
   return start.format("MMMM");
